refactor cmdline tests
