+++ pathoc
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
-from libpathod import cmdline
+
+from libpathod import pathoc_cmdline as cmdline
 
 if __name__ == "__main__":
     cmdline.go_pathoc()
