+++ pathod
@@ -1,5 +1,6 @@
 #!/usr/bin/env python
-from libpathod import cmdline
+
+from libpathod import pathod_cmdline as cmdline
 
 if __name__ == "__main__":
     cmdline.go_pathod()
