+++ test/test_pathod_cmdline.py
@@ -1,4 +1,4 @@
-from libpathod import cmdline
+from libpathod import pathod_cmdline as cmdline
 import tutils
 import cStringIO
 import mock
@@ -93,50 +93,6 @@ def test_pathod(perror):
     assert perror.called
     perror.reset_mock()
 
-
-@mock.patch("argparse.ArgumentParser.error")
-def test_pathoc(perror):
-    assert cmdline.args_pathoc(["pathoc", "foo.com", "get:/"])
-    s = cStringIO.StringIO()
-    tutils.raises(
-        SystemExit, cmdline.args_pathoc, [
-            "pathoc", "--show-uas"], s, s)
-
-    a = cmdline.args_pathoc(["pathoc", "foo.com:8888", "get:/"])
-    assert a.port == 8888
-
-    a = cmdline.args_pathoc(["pathoc", "foo.com:xxx", "get:/"])
-    assert perror.called
-    perror.reset_mock()
-
-    a = cmdline.args_pathoc(["pathoc", "-I", "10, 20", "foo.com:8888", "get:/"])
-    assert a.ignorecodes == [10, 20]
-
-    a = cmdline.args_pathoc(["pathoc", "-I", "xx, 20", "foo.com:8888", "get:/"])
-    assert perror.called
-    perror.reset_mock()
-
-    a = cmdline.args_pathoc(["pathoc", "-c", "foo:10", "foo.com:8888", "get:/"])
-    assert a.connect_to == ["foo", 10]
-
-    a = cmdline.args_pathoc(["pathoc", "-c", "foo", "foo.com:8888", "get:/"])
-    assert perror.called
-    perror.reset_mock()
-
-    a = cmdline.args_pathoc(
-        ["pathoc", "-c", "foo:bar", "foo.com:8888", "get:/"])
-    assert perror.called
-    perror.reset_mock()
-
-    a = cmdline.args_pathoc(
-        [
-            "pathoc",
-            "foo.com:8888",
-            tutils.test_data.path("data/request")
-        ]
-    )
-    assert len(list(a.requests)) == 1
-
     a = cmdline.args_pathod(
         [
             "pathod",
@@ -146,10 +102,3 @@ def test_pathoc(perror):
     )
     assert perror.called
     perror.reset_mock()
-
-    tutils.raises(
-        SystemExit,
-        cmdline.args_pathoc,
-        ["pathoc", "foo.com", "invalid"],
-        s, s
-    )
